requirements.txt
pandas
requests==2.31.0
psycopg2-binary
python-dotenv
sqlalchemy
